+++ src/utils/formatDate.ts
@@ -1,0 +1,23 @@
+/**
+ *
+ * @param dateString
+ * @returns 6월 4일
+ */
+export const formatDate = (dateString: string) => {
+  const date = new Date(dateString);
+  const now = new Date();
+  const diffInHours = Math.floor((now.getTime() - date.getTime()) / (1000 * 60 * 60));
+
+  if (diffInHours < 1) {
+    return '방금 전';
+  } else if (diffInHours < 24) {
+    return `${diffInHours}시간 전`;
+  } else if (diffInHours < 24 * 7) {
+    return `${Math.floor(diffInHours / 24)}일 전`;
+  } else {
+    return date.toLocaleDateString('ko-KR', {
+      month: 'short',
+      day: 'numeric',
+    });
+  }
+};
